deprecate RANCID migration page
The migration tool is needed once, and should not be part of the web
frontend if no one maintains it.
I also think there is no need for such a tool anymore - oxidized has
been around for a few years now.
The RANCID migration page will still be present in this major release,
but will be removed in a future major release if noone opens and issue
and steps forward to maintain it.