Fix group name containing "/"
I've got the use-case that the group name containes "/", which is used
to separate the city from the street in the git repository:
Stuttgart/Baerenstrasse/SW-42
Stuttgart/Baerenstrasse/R-42
Goldenstedt-Bockhorn/Flamingostrasse/SW-12
Goldenstedt-Bockhorn/Flamingostrasse/R-12

This commit fixes groups containing "/" in oxidized web and adds
multiple tests to validate this use-case and to be sure the change
does not break anything. node_spec.rb gets prety long, maybe we have to
split it in the future.

More changes in the tests:
- spec/spec_helper was improved to get the execptions in the output of
minitest (and not in last_response.body) - ENV['APP_ENV'] = 'test'
- we require oxidized in order to use its exception classes, so we do
not need to set the Oxidized::VERSION
- we do not follow the redirection in get '/', as we do not want to test
/nodes, but only the redirection from '/'. It is enought to test
last_response.location, http://example.com/ is the default hostname in
rack/test.