Fix /node/version/view (HTML-Encoded char in text and json)
/node/version/view should not encode HTML special characters when
returning the configuration in text or json format.

This commit closes #261

I've added tests reproducing the issue to confirm it has been fixed.

node_spec.rb was optimized, as we always need @nodes to be a mock.

I've also added a link (button) to the configuration in raw (text)
format. I didn't add a link to the configuration in json format because
I don't find it very useful...